Move DetermineLicenceIssues service
